refactor: delete store_backend_anonymizer
